pyproject.toml
[tool.poetry]
name = "multilingual-drc-news-chatbot"
version = "0.1.0"
description = "A project to build a question answering system using a congo-drc-new-corpus"
authors = ["espoir murhabazi <espoir.mur@gmail.com>", "merci jacob <mercihabam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Scrapy = "2.6.1"
alabaster = "0.7.12"
attrs = "21.4.0"
Automat = "20.2.0"
awscli = "1.22.87"
Babel = "2.9.1"
botocore = "1.24.32"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.12"
click = "8.1.2"
colorama = "0.4.3"
constantly = "15.1.0"
coverage = "6.3.2"
cryptography = "36.0.2"
cssselect = "1.1.0"
docutils = "0.15.2"
filelock = "3.6.0"
flake8 = "4.0.1"
greenlet = "1.1.2"
hyperlink = "21.0.0"
idna = "2.5"
imagesize = "1.3.0"
importlib-metadata = "4.2.0"
incremental = "21.3.0"
itemadapter = "0.5.0"
itemloaders = "1.0.4"
Jinja2 = "3.1.1"
jmespath = "1.0.0"
lxml = "4.8.0"
MarkupSafe = "2.1.1"
mccabe = "0.6.1"
packaging = "21.3"
parsel = "1.6.0"
Protego = "0.2.1"
psycopg2 = "2.9.3"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycodestyle = "2.8.0"
pycparser = "2.21"
PyDispatcher = "2.0.5"
pyflakes = "2.4.0"
Pygments = "2.11.2"
pyOpenSSL = "22.0.0"
pyparsing = "3.0.7"
python-dateutil = "2.7.3"
python-decouple = "3.6"
python-dotenv = "0.20.0"
pytz = "2017.2"
PyYAML = "5.4.1"
queuelib = "1.6.2"
requests = "2.23"
requests-file = "1.5.1"
rsa = "4.7.2"
s3transfer = "0.5.2"
service-identity = "21.1.0"
snowballstemmer = "2.2.0"
Sphinx = "4.3.2"
sphinxcontrib-applehelp = "1.0.2"
sphinxcontrib-devhelp = "1.0.2"
sphinxcontrib-htmlhelp = "2.0.0"
sphinxcontrib-jsmath = "1.0.1"
sphinxcontrib-qthelp = "1.0.3"
sphinxcontrib-serializinghtml = "1.1.5"
SQLAlchemy = "1.2.15"
tldextract = "3.2.0"
Twisted = "22.2.0"
typing-extensions = "4.1.1"
w3lib = "1.22.0"
zipp = "3.8.0"
"zope.interface" = "5.4.0"
scrapyd = "^1.3.0"
scrapyd-client = {git = "https://github.com/scrapy/scrapyd-client.git"}
python-scrapyd-api = "^2.1.2"
ipykernel = "^6.13.0"
scrapydweb = {git = "https://github.com/my8100/scrapydweb.git"}
Werkzeug = "2.0.3"
logparser = {git = "https://github.com/my8100/logparser.git"}
pylint = "^2.13.9"

[tool.poetry.dev-dependencies]
pycodestyle = "^2.8.0"
ipykernel = "^6.13.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
